requirements.txt
streamlit
python-dotenv
langchain
openai
langchain-community
langchain-openai
python-dotenv
